Added GHC 6.6.
note: the old GHC 6.4(.2?) version is still there. I reused the bootstrap-version to bootstrap GHC-6.6. And some other packages depend on the old GHC 
version as well.

note: only use this package for building other packages. If you install it as an end-user, you'll only be able to use the default GHC libraries, 
because other libraries are only privately registered to this GHC version during a nix-build by hooks (which are not executed when you run GHC 
yourself as an end-user).

Consequently, also added a newer version of uulib and uuagc.


git-svn-id: https://svn.nixos.org/repos/nix/nixpkgs/trunk@7346 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb