Fixed typo, added function add_to_var
The function for use in setup hooks. The syntax is very simple:

add_to_var varName needDir addDir prefix

If the directory $prefix$needDir exists, directory $prefix$addDir is added to
variable named varName (using PATH_DELIMITER as delimiter). If addDir is "", it
defaults to needDir (this is a most common case).


git-svn-id: https://svn.nixos.org/repos/nix/nixpkgs/trunk@9496 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb