git-git: change ${name} to reflect that this is not a pristine version
The git-git expression promises to build the current git HEAD, but in
fact it builds git HEAD plus a fairly extensive, undocumented, untested
patch that is apparently not part of the repository. By changing the
name of the installed package to 'git-git-with-glob-patch', we ensure
that users aren't mislead about what exactly they'll get when installing
this package. Technically, GIT-VERSION-FILE should be patched so that
the installed git binaries reflect this change too when called with
'--version', but the current solution is probably good enough.

git-svn-id: https://svn.nixos.org/repos/nix/nixpkgs/trunk@12506 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb